requirements.txt
Django==2.1
django-model-utils==3.1.2
django-extensions==2.1.0
Pillow==5.2.0
psycopg2==2.7.5
djangorestframework==3.8.2
Markdown==2.6.11
django-filter==2.0.0
pylint-django==2.0
django-address==0.2.0